+++ QSymbolicsBase/test/runtests.jl
@@ -1,6 +1,0 @@
-using QSymbolicsBase
-using Test
-
-@testset "QSymbolicsBase.jl" begin
-    # Write your tests here.
-end