+++ QSymbolicsClifford/test/runtests.jl
@@ -1,6 +1,0 @@
-using QSymbolicsClifford
-using Test
-
-@testset "QSymbolicsClifford.jl" begin
-    # Write your tests here.
-end