+++ QSymbolicsOptics/test/runtests.jl
@@ -1,6 +1,0 @@
-using QSymbolicsOptics
-using Test
-
-@testset "QSymbolicsOptics.jl" begin
-    # Write your tests here.
-end