+++ test/test_doctests.jl
@@ -1,9 +1,12 @@
 using Documenter
+using QuantumSymbolics
+using QuantumOptics
+using QuantumClifford
 
 function doctests()
     @testset "Doctests" begin
-        DocMeta.setdocmeta!(QuantumClifford, :DocTestSetup, :(using QuantumClifford); recursive=true)
-        doctest(QuantumClifford)
+        DocMeta.setdocmeta!(QuantumSymbolics, :DocTestSetup, :(using QuantumSymbolics, QuantumOptics, QuantumClifford); recursive=true)
+        doctest(QuantumSymbolics)
     end
 end