+++ test/test_qo.jl
@@ -1,7 +1,8 @@
 using Test
 using QuantumOptics
 using QuantumSymbolics
-using QSymbolicsOptics: LazyPrePost
+#using QuantumOpticsExt: LazyPrePost
+const LazyPrePost = Base.get_extension(QuantumSymbolics, :QuantumOpticsExt).LazyPrePost
 
 bs = GenericBasis(2),GenericBasis(2)
 op0 = Operator(bs...,rand(2,2))